UPSTREAM: soc/amd: Move root complex SSDT TOM1/TOM2 generation function
This will also be used for cezanne. Stoney also has a similar function,
but it hard codes the scope path. I didn't have a device setup to test
if switching to this function was a no-op. So I left it.

TOM2 isn't used by any ASL, so we could remove it later.

BUG=none
BRANCH=none
TEST=none

Change-Id: I73e0e6448fd896f4d6cdc485983f1b915698f3f5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 58a8ad1661ab0fd869bcbc955010717c011951b0
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Change-Id: I7c8f476a7735fea61a3244b97988e3ead3b42e79
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50892
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2711893
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>