UPSTREAM: cpu/intel/fit: Add the FIT table as a separate CBFS file
With CBnT a digest needs to be made of the IBB, Initial BootBlock, in
this case the bootblock. After that a pointer to the BPM, Boot Policy
Manifest, containing the IBB digest needs to be added to the FIT
table.

If the fit table is inside the IBB, updating it with a pointer to the
BPM, would make the digest invalid.
The proper solution is to move the FIT table out of the bootblock.

The FIT table itself does not need to be covered by the digest as it
just contains pointers to structures that can by verified by the
hardware itself, such as microcode and ACMs (Authenticated Code
Modules).

BUG=none
BRANCH=none
TEST=none

Change-Id: Ibb2fade56b26459818a1ddf812531ab56e98d446
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: eeacd8349c1bf208acf393e10e72a6da413a67ef
Original-Change-Id: I352e11d5f7717147a877be16a87e9ae35ae14856
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/50926
Original-Reviewed-by: Patrick Rudolph <patrick.rudolph@9elements.com>
Original-Reviewed-by: Christian Walter <christian.walter@9elements.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2775635
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>