UPSTREAM: rk3399: sdram: Remove obs_err variable to avoid uninitialized use
CB:50863 refactored the data_training() function to split out read gate
training into a separate function, but in the course of this forgot to
correctly initialize the local obs_err varible in the new function to 0.
This means that it will be used uninitialized, and when it happens to be
non-zero it makes the training process fail. Due to the convoluted
control flow in the function, it seems that GCC's static analyzer
couldn't pick up on this uninitialized use.

The whole variable is unnecessary anyway, all it's used for is to force
the function to return two lines below without doing anything with
side-effects in between. This patch removes the variable and simplifies
the code in all three training functions to avoid this uninitialized use
issue and make everything a bit more readable. (Also restore the
original pre-clang-format continuation line intendations for more
readability.)

BUG=none
BRANCH=none
TEST=none

Change-Id: I006d2d1d8d9db2bc864687c1466d18c9b8393905
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7f0d3de251f60dd71f48f87457c3057ec7a71148
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia475d64c06f2ec1bf9295742d173ce66717b821c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51079
Original-Reviewed-by: Moritz Fischer <moritzf@google.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2722218
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>