UPSTREAM: soc/intel/xeon_sp/cpx: Set the MRC "cold boot required" status bit
If bit 0 of byte 0x47 is set FSP will perform full memory training
even if previously saved data is supplied.

Up to and including FSP 2021 WW01 it was reset internally at the end
of PostMemoryInit. Starting with WW03 this is no longer the case and
Intel advised that this bit should be reset externally if valid MRC
data is present.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 95059b705500f4a7061bb3e8ea1e2fd01a3abde1
Original-Change-Id: I9c4191d2fa2e0203b3464dcf40d845ede5f14c6b
Original-Signed-off-by: Deomid "rojer" Ryabkov <rojer9@fb.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51230
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Change-Id: I97e69403f18f6ddee813f15f26389d2e20ea6624
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2752809
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>