UPSTREAM: soc/amd/picasso: Fix PSP_SHAREDMEM_BASE
PSP_SHAREDMEM_BASE made the assumption that _psp_sharedmem_dram would
only match once. With CB:49332 there are now two symbols, and it was
grabbing the wrong one.

This change makes it so we match the exact symbol. It also switches to
using awk to simplify the code.

The bootblock.elf target that is added to the list of prerequisites also
creates the bootblock.map file that gets used to extract the base
address of the _psp_sharedmem_dram symbol.

BUG=b:181354692
TEST=Boot zork past bootblock

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 42c5b010b6b2d8da75322fc7044fe77edf6a46f8
Original-Fixes: 82d16b150ce3 ("memlayout: Store region sizes as separate symbols")
Original-Suggested-by: Julius Werner <jwerner@chromium.org>
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I79675bd73f964282b54bca858830e26de64037c7
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51300
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Change-Id: I3c423344d169935022acf561dcd048703789f026
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2753188
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>