UPSTREAM: mb/google/brya: Reorganize flashmap
Intel ADL-P supports an additional memory-mapped 16MiB window into the
platform SPI flash. Support for this window already exists at the SoC
level, so all that is needed is to properly organize the flash map to
take advantage of this. FW_SECTION_A moves down to the bottom of the
available space in the lower 16MiB half, and FW_SECTION_B moves to the
bottom of the top 16MiB half. RW_LEGACY is squashed down to 2M.

BUG=b:182088676
TEST=build and boot to OS from FW_MAIN_A

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 659a591aa70e2df86c4ec576b8015ac78cd1ef03
Original-Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Change-Id: I60483b7e638c0a7e41f1f7e2b5503ae02e9906bd
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51345
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: EricR Lai <ericr_lai@compal.corp-partner.google.com>
Change-Id: I7b7643dd85357ec07c476aaf2e52828a7d322a73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2752823
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>