UPSTREAM: mb/google/dedede/var/madoo: Fix DPTF passive and critical policies
Some of the temperature sensors defined in baseboard do not exist in
madoo. With the format the DPTF policies are defined in madoo, all the
entries from the baseboard are included and then the overrides applied.
This causes the non-existent DPTF devices to be exported in the ACPI
table and in turn OS reading invalid temperatures. Fix the format for
DPTF passive and critical policies.

BUG=b:182513022
BRANCH=dedede
TEST=Build and boot to OS in madoo. Ensure that the DPTF entries look
correct in both static.c and SSDT tables i.e. passive and critical
policies for applicable devices only are present.

Change-Id: I0d7c8c05f3d7c6fbca3901afe8777c344f936bf8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ab8961e7be8ec6da5ad2d0dcaf6402bef419c210
Original-Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Original-Change-Id: Idc5d0b357d61b9346b4d20ec8322b124c9655b4c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51456
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Evan Green <evgreen@chromium.org>
Original-Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Original-Reviewed-by: Ben Kao <ben.kao@intel.com>
Original-Reviewed-by: Sumeet R Pawnikar <sumeet.r.pawnikar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2761378
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>