UPSTREAM: mb/intel/shadowwmountain: Enable CSE Lite SKU for shadowmountain
During the initial phases, the development and validation teams have to
deal with both Consumer SKU and Lite SKU firmware. Having the support for
CSE Lite enabled by default in coreboot helps in integrating both the SKUs.
With this we only have to interchange the CSE region in the full BIOS image
without having to worry about Kconfigs. Eases the build and integration
flow.

TEST=Verified build for Shadowmountain

Change-Id: I4dc6fde57fe9a2710c30e2357bc66f57cf7a40ff
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1ea938ca5e522cf0fbf93cbc8362b1ae58ec070f
Original-Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.com>
Original-Change-Id: I2ebf4da1b8c1df2e9c43b6e3bb688a9f8db652d3
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51496
Original-Reviewed-by: Maulik V Vaghela <maulik.v.vaghela@intel.com>
Original-Reviewed-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2790240
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>