UPSTREAM: mb/intel/adlrvp: Enable CSE Lite SKU
During the initial phases, the development and validation teams have to
deal with both Consumer SKU and Lite SKU firmware. Having the support for
CSE Lite enabled by default in coreboot helps in integrating both the SKUs.
With this we only have to interchange the CSE region in the full BIOS image
without having to worry about Kconfigs. Eases the build and integration
flow.

TEST= Built and booted on ADL-P LP4 RVP

Change-Id: Ifaacbdff95012fc8f0b6c6551bf81b5154cb6562
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d1150fd6d76e76fd1ccf8ede0ca96de0c404073c
Original-Signed-off-by: Sridhar Siricilla <sridhar.siricilla@intel.com>
Original-Change-Id: Ia92c7b71c69a23104ace9fc53fd39f01120fa751
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51567
Original-Reviewed-by: Rizwan Qureshi <rizwan.qureshi@intel.com>
Original-Reviewed-by: Maulik V Vaghela <maulik.v.vaghela@intel.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2790239
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>