UPSTREAM: security: vboot: Clarify PCR extension algorithms/sizes
The PCR algorithms used for vboot are frequently causing confusion (e.g.
see CB:35645) because depending on the circumstances sometimes a
(zero-extended) SHA1 value is interpreted as a SHA256, and sometimes a
SHA256 is interpreted as a SHA1. We can't really "fix" anything here
because the resulting digests are hardcoded in many generations of
Chromebooks, but we can document and isolate it better to reduce
confusion. This patch adds an explanatory comment and fixes both
algorithms and size passed into the lower-level TPM APIs to their actual
values (whereas it previously still relied on the TPM 1.2 TSS not
checking the algorithm type, and the TPM 2.0 TSS only using the size
value for the TCPA log and not the actual TPM operation).

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 74a0fad8a10ba401ce538fae9a4574e57fdff78d
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ib0b6ecb8c7e9a405ae966f1049158f1d3820f7e2
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51720
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Andrey Pronin <apronin@chromium.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Change-Id: I2f735afd57294ffbfaf89cf79a795fab4d160336
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2798512
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>