UPSTREAM: mb/google/dedede/var/boten: Fix DPTF passive and critical policies
Thermal sensor2 defined in baseboard do not exist in boten.
With the format the DPTF policies are defined in boten, all the entries
from the baseboard are included and then the overrides applied.
This causes the non-existent DPTF devices to be exported in the ACPI table
and in turn OS reading invalid temperatures. Fix the format for
DPTF passive and critical policies.

BUG=None
BRANCH=dedede
TEST=Build and boot to OS in boten. Ensure that the DPTF entries look
correct in both static.c and SSDT tables i.e. passive and critical
policies for applicable devices only are present.

Change-Id: I69a73b3e09b4e69861517253f8ee6ccba8567fd5
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 498f1285d2b7227c0f0e52ad5ed43c249412d907
Original-Change-Id: I63c781e0a439f1e7a3525fa7cf290fa9300cb066
Original-Signed-off-by: Stanley Wu <stanley1.wu@lcfc.corp-partner.google.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51722
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Karthik Ramasubramanian <kramasub@google.com>
Original-Reviewed-by: Ben Kao <ben.kao@intel.com>
Original-Reviewed-by: Sumeet R Pawnikar <sumeet.r.pawnikar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2790260
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>