UPSTREAM: soc/amd: add DISABLE_KEYBOARD_RESET_PIN option
The KBRST_L pin will cause a reset when driven or pulled low even when
the GPIO mux is set to GPIO and not native function. So when you want to
use that pin as general purpose output the keyboard reset input
functionality needs to be disabled by selecting this option in the
board's Kconfig file to avoid causing a reset by writing a 0 to the
output level bit when it's configured as an output.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 27b295b98b65e05635ee5a40967adb0f9ca79afa
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Suggested-by: Furquan Shaikh <furquan@google.com>
Original-Change-Id: I517ad551db9321f26afdba15d97ddb61be1f7d51
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51757
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Kangheui Won <khwon@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: Ifb8420b8e73e7a642b221dcd0a08ac593cf3dd3e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2795713
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>