UPSTREAM: soc/amd/cezanne,picasso: rename fsp_params.c to fsp_s_params.c
This file populates the UPD-S data structure that gets passed to the
FSP-S, so add that s part to make it a bit clearer which FSP parameters
it'll set up. This is also a preparation to add a fsp_m_params.c file in
the following patches.

BUG=none
BRANCH=none
TEST=none

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: 793f3717b4fd430f03455bdbbd071d5f7d104b57
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I53786df0909055e66eac675b5580909b7960944f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51833
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Change-Id: I608c12be9f440884cb2cd3abb76ab217842115a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2795715
Reviewed-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>