UPSTREAM: nb/intel/common: Replace `_bar_clrsetbits_impl` macro
This macro contains a cast on the and-mask, which can suppress actual
type overflow issues. Replace it with wrapper functions around the
existing macros in device/mmio.h which still contain a type cast, but
it is a non-issue because the wrapper functions now allow compilers to
check for overflows.

BUG=none
BRANCH=none
TEST=none

Change-Id: I087dd8830b4cb961da33e694f0a84766bba0f1af
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 54c4ecb9f26d5913f92d3568093d9cad37b8aab6
Original-Change-Id: I975bf8152fc961767f0292bff4a03aecd8c65f56
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/51886
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2867069
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>