UPSTREAM: decompressor: Add CBFS_VERIFICATION support
CBFS_VERIFICATION requires the CBFS metadata hash anchor to be linked
into an uncompressed stage, but for platforms using COMPRESS_BOOTBLOCK,
this is only the decompressor stage. The first CBFS accesses are made in
the bootblock stage after decompression, so if we want to make
CBFS_VERIFICATION work on those platforms, we have to pass the metadata
hash anchor from the decompressor into the bootblock. This patch does
just that. (Note that this relies on the decompressor data remaining
valid in memory for as long as the metadata hash anchor is needed. This
is always true even for OVERLAP_DECOMPRESSOR_ROMSTAGE() situations
because the FMAP and CBFS metadata necessarily need to have finished
verification before a new stage could be loaded.)

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie8b651b32c1874b0275f9af13af67620475f02d1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6296ca8ad9137a12094ad5d49d6f937dcf502105
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I2e6d7384cfb8339a24369eb6c01fc12f911c974e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52085
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2807684
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>