UPSTREAM: verstage: Add debug print when returning from verstage
RETURN_FROM_VERSTAGE is a somewhat tricky construct that we don't
normally do otherwise in coreboot. While it works remarkably well in
general, new development can lead to unintentional interactions with
confusing results. This patch adds a debug print to the verstage right
before returning to the bootblock so that it's obvious this happens,
because otherwise in some cases the last printout in the verstage is
about some TPM commands which can be misleading when execution hangs
after that point.

BUG=none
BRANCH=none
TEST=none

Change-Id: I2e80166a6b871c0c0984e072f39adb4a8cc99f57
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 5a090bf1257b6a5464ac21527d350b4aa13097ab
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: I9ca68a32d7a50c95d9a6948d35816fee583611bc
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/52086
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2807685
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>